backend/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.1
django-cors-headers==3.13.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Pillow==9.2.0
pycodestyle==2.9.1
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
